Improve GDI+ graphics driver
The GDI+ graphics driver maintains a Gdiplus::Graphics object
and uses it to draw, and rebuild it when necessary, that is,
when any change is done to the underlying HDC.
GDI+ is not used for horizontal and vertical lines
because GDI is more convenient to control adjacency of lines.
GDI+ is not used for text nor for images because GDI does
it very well.
GDI+ is used for everything that benefits from antialiasing.